:hammer: feat: BindingAdapter 구현, 분리
- View Component 별로 분리

Issue #121
